Handle players leaving in GameStart state
- If a player leaves ensure the correct action takes place
- Change the state back to OnePlayer if the player count ever reaches 1

Add tests for changing state on player leaves

If we remove the cap on two players, we should add an additional test for the condition (which checks that there is only one player left before changing back to `OnePlayer` state.
We should also add tests for the number of players on action 'playerLeaves'.