Remove 2 player limit
We were previously switching to GameStart when two players had joined the game. Instead we are displaying a button on two players joining to make 'playerClicksStart' a user triggered event to transition to 'GameStart'

Add a new test that more than 2 players can join, and that 'GameStart' state only exits if one player is left.