Add `.node-version`
By default, Bun will respect shebang lines in CLIs that indicate that
they should be executed with Node. We could force the use of Bun in CLIs
(like `eslint` and `stylelint`) using `--bun`, but this will sometimes
lead to unexpected behaviour. Playwright doesn't seem to work very well,
for instance

Meanwhile, using `nodenv` (which is fairly standard at dxw) without
specifying a `.node-version` can lead to errors like the following when
running CLIs that try to use it:

```
nodenv: node: command not found

The `node' command exists in these Node versions:
  16.17.0
  20.9.0
  20.10.0
  20.11.0
```

This adds a `.node-version` to fix this issue

See:
- https://bun.sh/docs/cli/run#bun
- https://bun.sh/docs/cli/bunx#shebangs