Add Biome
Biome is a linter and formatter built in Rust (fast). We're going to try
this as an alternative to ESLint or Standard. It has an opinionated
default config, but we can extend it if we like. I've configured it so
that [all rules are turned
on](https://biomejs.dev/reference/configuration/#linterrulesall), except
the more experimental [nursery
rules](https://biomejs.dev/reference/configuration/#linterrulesgroup)

Using an exact version of the package is recommended: https://biomejs.dev/internals/versioning

The state names are converted to regular camelcase here. There appears
to be no particular default/convention for this, so following Biome's
default seems the most sensible:
https://stately.ai/blog/2024-01-23-state-machines-whats-in-a-name#what-style-should-that-name-be-in