WIP render timer on page
This is currently rendering "15" on the page (for the initial timer
number). We need to decrease it by 1 each second. This is currently
separate from the XState countdown timer, but there might be a way to
read this value in future. This is a first step.

We've spotted a bug where this is currently rendering multiple times,
but we'll solve that soon.