Add setup function
After reading the [XState docs on
TS](https://stately.ai/docs/typescript#dynamic-parameters) I realised
that our current way of handling actions and guards doesn't allow for
typesafety.
"It is strongly recommended to use dynamic parameters instead of
directly accessing the event object whenever possible for improved type
safety and reusability.". This means we need to use the params function
when using a type rather than grabbing the data we need from the
context/event.
We should do this for all our machines but I thought I would keeps this
isolated for now in order to allow for quicker merging.
We also need to use the 'setup' function and chain 'createMachine' from
that to ensure full type safety... but this means we can remove the
typegen. More info in the docs here: https://stately.ai/docs/setup