WIP initialise turn machine outside of round machine
Rather than invoke the turn machine directly in the round machine we now
use the round model class as an intermediary to initiate the round
machine. Next one the round is in the 'finished' state we will need to
pass the received answers back from the turn to the round machine.