Remove card reorder on selection
While it's useful to be able to see all your selected colours in one
place, the way the reorder is currently implemented means that the
remaining unselected colours jump around a little (specifically any
colours that were to the right of the one just selected). If trying to
select answers quickly, this could lead to misclicks/frustration, so
we're removing the reorder for now. We can come up with an alternative
way of quickly seeing all selected colours later

Co-authored-by: Liz Daly <50752284+lookupdaily@users.noreply.github.com>