Move `submitAnswer` to utils
Here I also added `emitAnswersPost`, since it felt wrong to pass the
socket to the DOM manipulation utils. Instead, we pass a callback to
which it can pass the colours, and then the index will handle the
emission from the socket. Later we might want to move some of the socket
emission/message receipt logic into another separate file, but this is
out of scope for now

This also renames `submitAnswers` to `submitAnswer`, reflecting the
non-pluralisation of the type