Populate HTML elements on connection
When the DOM content is loaded we populate the HTML elements that we
want to use.
This works because we don't add or remove elements at the moment but
show and hide them.
I noticed in the files in the domManipulationUtils we were having to
populate the elements or check that they're populate every time their
used. This makes the functions more complicated than they need to be.