Watch bundled files
Whilst we were watching the serverside changes we weren't watching for
changes in the bundled (clientside) TS. This adds a new compile:watch
command and then uses it in the start:watch command. As compile:watch
never finishes (as it keeps watching) we need to run the two scripts
concurrently using a single '&'.