Add SocketIO package to externals
This package is imported via index.html and we are just using the types
here. But if we don't tell the bundler that it's externally imported it
then it will add it to the bundled code making a bigger and harder to
scan bundle.
Docs: https://bun.sh/docs/bundler#external