Check for a winner at turn end
We still need add the tiebreaker functionality if there are multiple
players with 10+ points, but for now we're just checking for a clear
winner and ending the round if there is one

It took a while to find a way to make this work. I tested it with the
minimum winning score set to 2 points. Eventually I found a way to do it
by adding an intermediate `checkForWinner` state, but it would be nice
to be able to do it without that. The hard parts are always doing the
update scores action, and always doing it before the guard runs. We
aren't allowed to have transitions in the final state, so we can't use
guards there to return to the turn state