Render scores and bonus points
This is currently on every round state change. Since we have guards on
`checkForWinner`, we never actually 'enter' that state, so this will
just be on the start of a turn and at round end. If we add more states
to the round, we might end up updating it more than expected though