Add clientside code
This renders the time out and counts down when the event is received.
We add the setTimeoutId to the window so we can clean it up when we
remove the timer.
I noticed when running locally that the player that submits the first
answer sees the timer a small but noticable (subsecond) time before the
other user. Whilst this won't effect the fairness of the game it is not
ideal. It would be good to work out where this latency is coming from as
both the clients should receive the 'countdown:start' event at (almost)
the same time.
It will need a lot of work visually but this is a first pass.