Pass countdown method through to machine
This is a different approach to how we have done things in the past.
With the other state machines we send messages to the client via
subscribing to the state of the state machine and switching on it. This
could get messy as we are firing actions in two seperate places - the
'model' and the machine itself. XState allows the machine to pass in the
action functions when the machine is instantiated with the [machine's
`provide`
function](https://stately.ai/docs/actions#implementing-actions). This is
then used in the machine as an action. We have to declare it in the
machine definition but it is then overwriten in 'provide'. I think this
would be a good way to do similar side effects moving forward so am keen
for feedback on it.