Remove player:get event
Now the client knows what state the server is we no longer need to tell
it to get players. It just does it every time we enter the "empty",
"onePlayer" or "multiplePlayers" states because a transition to one of
these states indicates the players could have changed. If we are in the
"multiplePlayers" state and receive a playerJoins event we will still
transition to the "multiplePlayers" even though we are already in it so
we will still getPlayers even when there are 2+ players.