Style colour card section
This may make more sense in the colour-card.css file but as it was
styling the top level component I put it in the general css file.
I think it's fine that the colour-cards file doesn't use the CSS vars as
it's a seperate component, but it makes sense for all the components in
this file to.