Store session in socket.data
This will allow us to fetch the session later by accessing the
socket.data object in the server or client.
We no longer autoconnect but only connect once a user has submitted
their name. This prevents users being added to the session before they
have comitted to the game. This allows users to retain their gamestate
if their connection drops for any reason.
This is following the guidance outlined here:
https://socket.io/get-started/private-messaging-part-1/

Co-authored-by: Ynda Jas <ynda@dxw.com>
Co-authored-by: George Eaton <george@dxw.com>
Co-authored-by: Liz Daly <liz.daly@dxw.com>