Extract addPlayerHandler
We want this logic in two places - when a player clicks join game
(listening to the 'players:post') event and when restoring session of a
previously joined player. It makes sense to abstract it to a function. I
think it would be good if we extracted all event handlers like this to
improve readability.